Rename jwt_test.go to user_test.go, add product unit test and add
Changed the file jwt_test.go to user_test.go.

Added a unit test for adding products.
